+++ NetCoreRestApi/DataLayer.EF/Entities/ProductEntity.cs
@@ -6,6 +6,6 @@ namespace DataLayer.EF.Entities
     {
         public decimal Price { get; set; }
         public int AvailableCount { get; set; }
-        public IEnumerable<CategoryEntity> Categories { get; set; }
+        public ICollection<CategoryEntity> Categories { get; set; }
     }
 }
