+++ NetCoreRestApi/DataLayer/UnitOfWorks/CategoryUnitOfWork.cs
@@ -7,7 +7,7 @@ namespace DataLayer.UnitOfWorks
 {
     public class CategoryUnitOfWork : BaseUnitOfWork<CategoryModel, int>, ICategoryUnitOfWork
     {
-        public CategoryUnitOfWork(IUnitOfWorkDbContext dbContext, ICategoryRepository categoryRepository) :
+        public CategoryUnitOfWork(IUnitOfWorkContext dbContext, ICategoryRepository categoryRepository) :
             base(dbContext, categoryRepository)
         {          
         }
